requirements.txt
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2
idna==2.10
paramiko==2.7.2
pycparser==2.20
PyNaCl==1.4.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.11
